sprint-2/step-2 Реализован drag and drop
