+++ src/components/app/app.jsx
@@ -7,32 +7,17 @@ import BurgerIngredients from "../burger-ingredients/burger-ingredients";
 import appStyles from "./app.module.css";
 import { ingredientListCreator } from "../../utils/helper";
 import { IngredientsDataContext } from "../../services/ingredients-data-context";
-import {
-  applayIngredients,
-  ADD_CONSTRUCTOR_ITEM
-} from "../../services/actions/index.js";
+import { applayIngredients } from "../../services/actions/ingredients";
 import { DndProvider } from "react-dnd";
 import { HTML5Backend } from "react-dnd-html5-backend";
 
 function App() {
   const dispatch = useDispatch();
-  const { ingredients, ingredientsRequest, ingredientsFailed, constructorIngredients } = useSelector(state => state);
+  const { ingredients, ingredientsRequest, ingredientsFailed } = useSelector(state => state.ingredients);
 
   React.useEffect(() => {
     dispatch(applayIngredients());
   }, [dispatch]);
-
-  React.useEffect(() => {
-    if (constructorIngredients) {
-      constructorIngredients.map(item => {
-        dispatch({
-          type: ADD_CONSTRUCTOR_ITEM,
-          typeItem: item.type,
-          idType: item._id
-        })
-      })
-    }
-  }, [constructorIngredients, dispatch])
 
   return (
     <div className={appStyles.app}>
