+++ src/components/order-details/order-details.jsx
@@ -5,7 +5,7 @@ import orderDetailsStyles from "./order-details.module.css";
 import checkIcon from "../../images/check.svg"
 
 const OrderDetails = () => {
-    const { orderId } = useSelector(state => state);
+    const { orderId } = useSelector(state => state.order);
     return (
         <section className={orderDetailsStyles.wrapper}>
             <span className={`${orderDetailsStyles.orderId} text text_type_digits-large mb-8 mt-9`}>{orderId}</span>
