+++ src/services/actions/burgerConstructor.js
@@ -1,0 +1,2 @@
+export const ADD_CONSTRUCTOR_ITEM = "ADD_CONSTRUCTOR_ITEM";
+export const REMOVE_CONSTRUCTOR_ITEM = "REMOVE_CONSTRUCTOR_ITEM";
