+++ src/services/actions/modal.js
@@ -1,0 +1,5 @@
+export const SHOW_INGREDIENT_MODAL = "SHOW_INGREDIENT_MODAL";
+export const CLOSE_INGREDIENT_MODAL = "CLOSE_INGREDIENT_MODAL";
+
+export const SHOW_ORDER_MODAL = "SHOW_ORDER_MODAL";
+export const CLOSE_ORDER_MODAL = "CLOSE_ORDER_MODAL";
